+++ examples/adc.rs
@@ -38,8 +38,8 @@ fn main() -> ! {
         adc::Adc::adc1(dp.ADC1, &mut delay, ccdr.peripheral.ADC12, &ccdr.clocks).enable();
     adc1.set_resolution(adc::Resolution::SIXTEENBIT);
 
-    let mut adc1_channel_4 = pins.SEED_PIN_21.into_analog(); // Daisy Pod: POT_1
-    let mut adc1_channel_0 = pins.SEED_PIN_15.into_analog(); // Daisy Pod: POT_2
+    let mut adc1_channel_4 = pins.GPIO.PIN_21.into_analog(); // Daisy Pod: POT_1
+    let mut adc1_channel_0 = pins.GPIO.PIN_15.into_analog(); // Daisy Pod: POT_2
 
     // - led ------------------------------------------------------------------